# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Users",
    "version": "1.0",
    "description": "user project",
    "contact": {
      "name": "Viktoriia",
      "url": "https://github.com/HladkaViktoriia",
      "email": "vktr.gldk10.98@gmail.com"
    }
  },
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "users"
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "get users",
            "schema": {
              "$ref": "#/definitions/status_200"
            }
          },
          "400": {
            "description": "just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      },
      "post": {
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "parameters": [
          {
            "name": "userName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "uniqueItems": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          }
        ,
          {
            "name": "lastName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "formData",
            "default": "driver",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "default": "Vika@gmail.com",
            "required": true,
            "uniqueItems": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "123456",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "create user",
            "schema": {
              "$ref": "#/definitions/status_200"
            }
          },
          "400": {
            "description": "just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    },
    "/users/{id}": {
      "put": {
        "tags": [
          "user"
        ],
        "operationId": "updateUser",
        "parameters": [
          {
            "name": "userName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "uniqueItems": true,
            "type": "string"
          },
          {
            "name": "firstName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "default": "Vika",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "formData",
            "default": "driver",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "default": "Vika@gmail.com",
            "required": true,
            "uniqueItems": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "default": "123456",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "update user",
            "schema": {
              "$ref": "#/definitions/status_200"
            }
          },
          "400": {
            "description": "just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "user"
        ],
        "operationId": "deleteUser",
        "description": "This is how to delete a user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "user deleted",
            "schema": {
              "$ref": "#/definitions/status_200"
            }
          },
          "400": {
            "description": "just an error",
            "schema": {
              "$ref": "#/definitions/status_400"
            }
          }
        }
      }
    }
  },

  "definitions": {
    "status_200": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "success"
        },
        "code": {
          "type": "number",
          "example": 200
        }
      }
    },
    "status_400": {
      "type": "object",
      "properties": {
        "msg": {
          "type": "string",
          "example": "something went wrong"
        },
        "code": {
          "type": "number",
          "example": 400
        }
      }
    }
  }
}